Login form implemented

+++ src/AppBundle/Controller/NewsController.php
@@ -5,36 +5,48 @@ namespace AppBundle\Controller;
 use AppBundle\Entity\News;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;use Symfony\Component\HttpFoundation\Request;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
-/**
- * News controller.
- *
- * @Route("news")
- */
 class NewsController extends Controller
 {
     /**
-     * Lists all news entities.
+     * Lists all news .
      *
      * @Route("/", name="news_index")
      * @Method("GET")
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
-
-        $news = $em->getRepository('AppBundle:News')->findAll();
-
+        $news = $this->getLatestNews();
+        $newsImagesPath = $this->container->getParameter('news_images_path');
         return $this->render('news/index.html.twig', array(
             'news' => $news,
+            'newsImagesPath' => $newsImagesPath
+        ));
+    }
+
+    /**
+     * Lists all current users news .
+     *
+     * @Route("/my_news", name="my_news")
+     * @Method("GET")
+     */
+    public function myNewsAction()
+    {
+        $news = $this->getLatestNews($this->getUser());
+        $newsImagesPath = $this->container->getParameter('news_images_path');
+        return $this->render('news/index.html.twig', array(
+            'news' => $news,
+            'newsImagesPath' => $newsImagesPath
         ));
     }
 
     /**
      * Creates a new news entity.
      *
-     * @Route("/add", name="news_add")
+     * @Route("/news/add", name="news_add")
      * @Method({"GET", "POST"})
      */
     public function addAction(Request $request)
@@ -44,7 +56,7 @@ class NewsController extends Controller
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $news->setUserId(1);
+            $news->setUser($this->getUser());
             $em = $this->getDoctrine()->getManager();
             $em->persist($news);
             $em->flush($news);
@@ -59,18 +71,17 @@ class NewsController extends Controller
     }
 
     /**
-     * Finds and displays a news entity.
+     * Finds and displays a news .
      *
-     * @Route("/{id}", name="news_show")
+     * @Route("/news/details/{id}", name="news_show")
      * @Method("GET")
      */
     public function showAction(News $news)
     {
-        $deleteForm = $this->createDeleteForm($news);
-
+        $newsImagesPath = $this->container->getParameter('news_images_path');
         return $this->render('news/show.html.twig', array(
             'news' => $news,
-            'delete_form' => $deleteForm->createView(),
+            'newsImagesPath' => $newsImagesPath,
         ));
     }
 
@@ -78,44 +89,79 @@ class NewsController extends Controller
     /**
      * Deletes a news entity.
      *
-     * @Route("/delete/{id}", name="news_delete")
+     * @Route("/news/delete/{id}", name="news_delete")
      * @Method("GET")
      */
     public function deleteAction(Request $request, News $news)
     {
-//        $form = $this->createDeleteForm($news);
-//        $form->handleRequest($request);
-//
-//        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($news);
-            $em->flush($news);
-//        }
+            if($news->getUser()->getId() == $this->getUser()->getId()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($news);
+                $em->flush($news);
+                $this->addFlash(
+                    'error',
+                    'news deleted successfully .'
+                );
+            }
+            else {
+                $this->addFlash(
+                    'error',
+                    'You do not have permission to delete that news.'
+                );
+            }
+           return $this->redirectToRoute('news_index');
 
-        return $this->redirectToRoute('news_index');
     }
 
      /**
      * Download news .
      *
-     * @Route("/{id}/download", name="news_download")
+     * @Route("/news/{id}/download", name="news_download")
     * @Method({"GET", "POST"})
      */
     public function downloadAction(Request $request, News $news)
     {
-
-//        dump($request);die;
-
+        $newsImagesPath = $this->container->getParameter('news_images_path');
         $repository = $this->getDoctrine()
             ->getRepository('AppBundle:News');
         $pdf = $this->container->get("white_october.tcpdf")->create();
-        $repository->downloadPdf($news, $pdf,  $this->container->get('request')->server->get('DOCUMENT_ROOT'));
-
-
+        $repository->downloadPdf($news, $pdf,  $this->container->get('request')->server->get('DOCUMENT_ROOT'), $newsImagesPath);
 
         return $this->redirectToRoute('news_index');
     }
 
+    /**
+     * Rss news .
+     *
+     * @Route("/news/rss/", name="news_rss")
+     * @Method({"GET"})
+     */
+    public function rssAction(Request $request)
+    {
+        $news = $this->getLatestNews();
+        $baseUrl = $request->getScheme() . '://' . $request->getHttpHost();
+        $newsImagesPath = $this->container->getParameter('news_images_path');
+        $response = new Response($this->render('news/rss.xml.twig', array(
+            'news' => $news,
+            'baseUrl' => $baseUrl,
+            'newsImagesPath' => $newsImagesPath
+        )));
+        $response->headers->set('Content-Type', 'xml');
+        return $response;
+    }
+
+    public function getLatestNews($user  = null) {
+        $em = $this->getDoctrine()->getManager();
+        if($user) {
+            $conditions = array('user' => $user->getId());
+            return  $em->getRepository('AppBundle:News')->findBy($conditions, array('created'=> 'DESC'));
+        }
+        else {
+            $conditions = array();
+            return  $em->getRepository('AppBundle:News')->findBy($conditions, array('created'=> 'DESC'),$this->container->getParameter('page_limit') );
+        }
+
+    }
 
     /**
      * Creates a form to delete a news entity.
